requirements.txt
requests
beautifulsoup4
markdownify
python-dotenv
openai
tiktoken
boto3